Moved the elements folder in to the support folder
