+++ cypress/integration/canvas_test_spec.js
@@ -1,8 +1,8 @@
-import LeftNav from './elements/LeftNav'
-import Canvas from './elements/Canvas'
-import RightNav from './elements/RightNav'
-import BottomNav from './elements/BottomNav';
-import LearningLog from './elements/LearningLog';
+import LeftNav from '../support/elements/LeftNav'
+import Canvas from '../support/elements/Canvas'
+import RightNav from '../support/elements/RightNav'
+import BottomNav from '../support/elements/BottomNav';
+import LearningLog from '../support/elements/LearningLog';
 
 context('Test Canvas', function(){
     let leftNav = new LeftNav;
