+++ cypress/integration/group_chooser_spec.js
@@ -1,4 +1,4 @@
-import Header from './elements/Header'
+import Header from '../support/elements/Header'
 
 const header = new Header;
 const baseUrl = `${Cypress.config("baseUrl")}`;
