+++ cypress/integration/group_test_spec.js
@@ -1,5 +1,5 @@
-import LeftNav from './elements/LeftNav'
-import Canvas from './elements/Canvas'
+import LeftNav from '../support/elements/LeftNav'
+import Canvas from '../support/elements/Canvas'
 
 const leftNav = new LeftNav;
 const canvas = new Canvas;
