+++ cypress/integration/my_work_test_spec.js
@@ -1,6 +1,6 @@
-import Workspace from './elements/Workspace'
-import RightNav from './elements/RightNav'
-import Canvas from './elements/Canvas'
+import Workspace from '../support/elements/Workspace'
+import RightNav from '../support/elements/RightNav'
+import Canvas from '../support/elements/Canvas'
 
 describe('Test right nav tabs', function(){
