+++ cypress/integration/teacher_test_spec.js
@@ -1,9 +1,9 @@
-import Header from './elements/Header.js';
-import RightNav from './elements/RightNav';
-import LeftNav from './elements/LeftNav';
-import Canvas from './elements/Canvas';
-import Workspace from './elements/Workspace';
-import TeacherDashboard from './elements/TeacherDashboard';
+import Header from '../support/elements/Header.js';
+import RightNav from '../support/elements/RightNav';
+import LeftNav from '../support/elements/LeftNav';
+import Canvas from '../support/elements/Canvas';
+import Workspace from '../support/elements/Workspace';
+import TeacherDashboard from '../support/elements/TeacherDashboard';
 
 let qaClass = 10,
     qaOffering = 10,
