+++ cypress/support/elements/Canvas.js
