+++ cypress/support/elements/DrawToolTile.js
