+++ cypress/support/elements/GraphToolTile.js
